refactor : add Dao Interface
